Update patch set 1
Patch Set 1:

Lalit;

Yeah, this (QDSP5 stuff) is a really old problem, and we (and me personally) have been struggling with it for about 3 years now.
Considering qcom has published the code for the more recent interfaces (q6 and vidc), I suspect there are ownership/IP issues involved in the QDSP5 APIs that they can't solve. I'm sure this API hasn't been deprecated (7227A is an active platform and it also uses q5), it simply has never been published to the public.

Patch-set: 1